backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0  # Updated to support anyio 4.x (required by google-genai)
uvicorn[standard]>=0.30.0  # Updated to support anyio 4.x
python-multipart==0.0.6

# Google Cloud Services
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
# Vertex AI - using new google-genai SDK for Gemini 2.5 Pro
google-genai>=0.2.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Data Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Handling
aiofiles==23.2.1
PyPDF2>=3.0.0          # PDF text extraction
openpyxl>=3.1.0        # Excel xlsx reading
pdfplumber>=0.10.0     # Better PDF text extraction (tables)

# Machine Learning Core (Python 3.12 compatible)
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
spacy[cpu]>=3.7.0
tensorflow>=2.15.0
Pillow>=10.0.0

# Transformers for DistilBERT
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.26.0  # Required for PyTorch Trainer
tf-keras>=2.20.0  # Required for transformers compatibility with Keras 3

# Model Serialization
joblib>=1.3.0

# HTTP Client (for health checks and URL downloads)
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# Process Management (for Scratch services)
psutil>=5.9.0

# Production
gunicorn>=21.2.0